Review from level 6 snd ms audrienne
ms audrienne- spawning less close together make a code to help it have some distance in between as well as make jumping function work properly especially this close. Important that the elements are not just red and green and more specific. Try to make it more fair in terms of elements. She mentioned she enjoyed the ducking movement.

level 6- make the speed gradually move make the beginning a bit slower to make the functions easier to use. Make the mother at the beginning have a shoe constantly in her hand as she wont have time to collect the belt and make her at some time throw the shoe at the player. They mentioned i should use a fit to screen key frame in m code to make sure that the elements whenever the screen size is changed the layout will fix itself authomatically. They also added that it would be more ideal for me to use the screen as fullhd