update
Added points as well as high score. Made sure to make objects disapear once the player goes through them and gives him a point. Created the main menu to where the game will begin. Added a gameManager to control the beginning of the game as well as how the speed will appear once the game begins